feat(Board generation): Add a script which automatically generates the board.
It Generates the board according to its description in boards.json and correctly
sets the ids of the elements of the board.

The SVG version of the board (as generated by the script) is also added to
the project since it will not be generated many times. It needs correction.