feat(TurnPlayer, Main): Turn by turn implementation.
Creating a turn by turn game mechanism allowing only the player whose turn it is to play. For a player, playing while then give the hand to the next non null player (for less than 8 players). If a player is the only player in game, a message is displayed but he can still play alone. The first player to play is arbitrarily chosen.