fix(CreateGameRest, PlayRest, PossibleSquaresRest, PossibleSquaresLister): #434.
The problem was that when I was creating a game in CreateGame, I was
only storing the Match object in the session scope. When I needed the
deck and the board, I was creating a new GameFactory. So the board used
by the card to compute the possible movements had not players in it, so
no occupied squares.

I now store the GameFactory and get all the necessary items from it.

Need refactoring.