feat(protractor): enable protractor with firefox.
The bug which prevented the use of Firefox with protractor is solved.

I had to modify the way cards are selected to go around another bug with
Firefox and protractor. See
https://github.com/angular/protractor/issues/1896.

Tests on firefox only pass if all the element we interact with are
visible. In the current design, trumps may be hidden. Take care.