feat(game-ctrl, game-directives): Limit certain actions to the game master.
only the game master can add players, create the game or change the
state of the slots.

If a field is taken, nobody can change its state. WARNING: if the game
master sets a slot to taken, it is not updated in the database and
nobody can interract with it.