fix(create,api): allow the player to create a new game after the game over
When the player clicked on the "New Game" button of the game over popup,
he/she was redirected to the game creation page but the game popup
stayed on top and the player could not create a new game.

- When the player clicks on game over button, we resolve the promise and
close the popup.
- Flush the Wait.forId cache when initing a new game. Since the page
changed, the previous promise where resolved with unexisting elements.
- Add an init method to Api and Create so both object can be reset on
new game. The init method of Create is called when there is no game id
or if the registered game id doesn't match the save one for the
registered player id.