fix(history): init history on game page so we are sure it is setup
If the page is reloaded on the game page, the history may not be setup
until the player click on the player box. This may result in some
actions not being displayed. For instance, create a game, refresh, play
a card. Without this fix, it will not appear in the player box.