fix: prevent t-params binding to be lost when changing locale
There is a bug in aurelia-i18n that caused t-params to be lost when
locate is changed. This affects "It is the turn of". Update aurelia-i18n
to a patched version to avoid this problem.

Bug report: https://github.com/aurelia/i18n/issues/116
PR: https://github.com/aurelia/i18n/pull/130