fix(select-heroes): correct arrow size and position
Intead of waiting for promise which may not be resolved in time, we rely
on the fact that until an image is loaded its naturalWidth property is
equal to 0. If it is, we wait 500ms and call resize with the same
parameters.

We still need to iterate in order for the position to be correct.