fix(play,counter,notifications): prevent promise rejection on game creation
The onReconnectDefered promise will be resolved when the game is
created. We don't want to do anything with special action at this point.
To do so, we loosen the condition on special_action_name not to execute
the functions that notify a special action if special_action_name isn't
in the message.