ref(infos): don't rely on id style element
Now that we have one instance of aot-infos per aot-trump, the id
trumps-element-infos could be duplicated. This caused the trump popup
not to be updated. We inject the element instead. For clarity,
this.element is now the aot-infos element and this.infosDiv the element
we used to get with the id.