fix(trump): prevent the frame of the trumps to disapear
On some browsers (Chrome), when an affecting trump was removed from the
DOM, the gradients were not applied any more on the frames of the
trumps. This was caused by the fact that all the gradient shares the
same ids. When the affecting trump was removed, the browser considered
that the gradient was removed.

We prevent that by changing all the ids of the gradients when attaching
the component so they are specific to each component.