ref(gauge,trump): change how images in SVG are loaded
We used to load them with somethin like:
<pattern id="hero">
    <image></image>
</pattern>

And to use them with:
<ellipse style="fill: url(#hero)" />

But FireFox may not display the images (even if they are correctly
loaded according to the network tab of the dev tools) when the user has
a "slow" connection.

To prevent this, we directly use the image tag in the SVG.

Affected components: aot-gauge and aot-trump. In the case of aot-trump,
we also simplify the associated JS: we don't need to rewrite the id of
the pattern in order for it to be specific to an instance of the
component.