chore(lint): correctly lint SCSS files
Linting with `stylelint --syntax scss app/**/*.scss` in the stylelint
caused problems since it is platform dependent. It works on Windows but
not on Linux (eg in the bitbucket pipelines). In order for it to work on
Linux, we need to put the files in quotes:
`stylelint --syntax scss 'app/**/*.scss'`

But then it doesn't work on Windows.

To solve this, we move the linting into a custom nodejs script and we
update the npm stylelint script accordingly. It works on Linux and
Windows and shouldn't cause any trouble on other platforms.