fix(create): make sure we register observers after each activation
To do that, we register them in the init function instead of the
initPlayerInfos: the init function is called each time the view is
activated.

The fact that they were not registered on each activation caused a bug:
in the process of creating the game, /create is activated, then
deactivated to activate /create/<game_id> (which activates the create
view). So the observers were reregistered but never registered again. We
now make sure they are.