fix(game): display the transition popup and start the counter on 1 player games
If the last player of the game (all other disconnected) passed his/her
turn because of the counter, it was reset to 90 seconds but never
started. The transition popup was also never displayed. We fix this by
taking into account the number of turns: it will be increased by one in
this case.