fix(trump): correctly display player names when there is hole in the slots
The playerIndexes array contains the index of the players. When there is
a hole, it can be [0, 2]. We would then try to read the name in position
2 of the player names which was undefined.

To prevent this, we use a standard for loop to loop on both array at the
same time. This way the index we use to read the index and the name will
match and the name will correctly be displayed.