test: fix tests and running tests
We rely on nps to run jest through its CLI. It will allow us to pass
options to it more easily. We must rely on nps to correctly set an env
var.

We thus deprecate the test command from Aurelia CLI.

We also move Jest configuration into package.json. This was for some
reason required anyway because relying on aurelia.json caused issues
with assets imports.